pyproject.toml
[tool.poetry]
name = "hpe-agent-rag"
version = "0.1.0"
description = ""
authors = ["Shahid"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.6"
pydantic = "^2.10.5"
passlib = "^1.7.4"
boto3 = "^1.35.98"
pyyaml = "^6.0.2"
uvicorn = "^0.34.0"
python-jose = "^3.3.0"
bcrypt = "^4.2.1"
jwt = "^1.3.1"
langchain = "^0.3.14"
langchain-aws = "^0.2.11"
langgraph = "^0.2.64"
requests = "^2.32.3"
requests-aws4auth = "^1.3.1"
opensearch-py = "^2.8.0"
psycopg2-binary = "^2.9.10"
langchain-community = "^0.3.14"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
